+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Calva.
 
 ## [Unreleased]
 
+- Fix: [Trim leading whitespaces from string before using as inline result/error decoration](https://github.com/BetterThanTomorrow/calva/issues/2617)
+
 ## [2.0.470] - 2024-09-11
 
 - Bump deps.clj to v1.12.0.1479